finished up

+++ app/Http/Controllers/ListingController.php
@@ -36,6 +36,8 @@ class ListingController extends Controller
         if($request->hasFile('logo')){
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
         return Redirect::route('index')->with(['message' => 'Listing created successfully!']);
     }
@@ -43,6 +45,11 @@ class ListingController extends Controller
         return view('listings.edit', ['listing' => $listing]);
     }
     public function update(Request $request, Listing $listing){
+
+        if($listing->user_id != auth()->id()){
+            abort(403, 'unauthorized action');
+        }
+        
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -59,7 +66,14 @@ class ListingController extends Controller
         return back()->with(['message' => 'Listing updated successfully!']);
     }
     public function destroy(Listing $listing){
+        if($listing->user_id != auth()->id()){
+            abort(403, 'unauthorized action');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
