+++ routes/web.php
@@ -9,27 +9,31 @@ use App\Http\Controllers\ListingController;
 
 Route::get('/', [ListingController::class, 'index'])->name('index');
 
-Route::get('/listings/create', [ListingController::class, 'create']);
-
-Route::post('/listings', [ListingController::class, 'store']);
-
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
-
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
-
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
+
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
+
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
+
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
+
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 Route::post('/users', [UserController::class, 'store']);
 
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->name('logout')->middleware('auth');
 
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
+
+
 /*
 Route::get('/hello', function() {
     return response('<h1>Hello world</h1>')
